Opcodes File uploaded
uploaded a text file with the opcodes known so far.
MIPS can have up to 4 coprocessors. The most common two are cop0 for
exception handling and memory management and cop1 for the FPU their
opcodes are stored in binary as 0100xx where xx are used to tell the
coprocessor number.